Moved CardType into CardContent.ts
