+++ types/CardContent.ts
@@ -1,4 +1,15 @@
 /**
+ * Card Types
+ *
+ * @description Enumeration of card types supported by the CardGrid component.
+ */
+
+export enum CardType {
+  Service = 'service',
+  Review = 'review',
+}
+
+/**
  * Service Card Content Interface
  */
 export interface IServiceCardContent {
