--- types/CardTypes.ts
@@ -1,10 +1,0 @@
-/**
- * Card Types
- *
- * @description Enumeration of card types supported by the CardGrid component.
- */
-
-export enum CardType {
-  Service = 'service',
-  Review = 'review',
-}
